+++ sentry/src/main/java/io/sentry/IScopes.java
@@ -528,6 +528,16 @@ public interface IScopes {
   }
 
   /**
+   * Captures the profile chunk and enqueues it for sending to Sentry server.
+   *
+   * @param profileChunk the continuous profiling payload
+   * @return the profile chunk id
+   */
+  @ApiStatus.Internal
+  @NotNull
+  SentryId captureProfileChunk(final @NotNull ProfileChunk profileChunk);
+
+  /**
    * Creates a Transaction and returns the instance.
    *
    * @param transactionContexts the transaction contexts
@@ -581,6 +591,10 @@ public interface IScopes {
   ITransaction startTransaction(
       final @NotNull TransactionContext transactionContext,
       final @NotNull TransactionOptions transactionOptions);
+
+  void startProfiler();
+
+  void stopProfiler();
 
   /**
    * Associates {@link ISpan} and the transaction name with the {@link Throwable}. Used to determine